"10:20am. I slept in today. Let me do some chilling first, then breakfast, and after that I will start.
10:50am. https://www.reddit.com/r/ProgrammingLanguages/comments/ew6y8s/quest_for_a_language_sth_along_the_lines_of_ml_gc/
"[Spiral](https://github.com/mrakgr/The-Spiral-Language) definitely fits the bill for what you are describing. Back in 2018 I used it to implement a ML library that allowed passing of lambdas across the language boundaries to the GPU backend, monadic computation on the GPU, struct-of-array tensors. Because it has exactly what you need to take full advantage of inlining, writing both Cuda kernels and host code in a high level functional style is not only possible, but encouraged.

The intended use case is when you have a runtime and a GC and you want to take advantage of that to write code in a high level manner, but you also have some device like a GPU that needs absolute control and cannot use GC or even heap allocation.

Right now, the language is not in a good shape though. I did a lot of work on v0.09, but I've made various errors that made the compilation times too long. The type inference is done bottom-up, so it is similar to dynamic languages where you need to run the code to typecheck it (even though this is done at compile time.) This is a problem as top-down is much more ergonomic since it gives immediate feedback.

Back then I did not know how to deal with this, but I do now and v0.2 is going to be what Spiral should have been from the start. Also since it is 2020, rather than aiming for GPUs, it is finally time to go for [neuromorphic chips](https://youtu.be/jhQgElvtb1s)."

Self promotion is fun.

12:20pm. Done with breakfast. The next come the chores.

1:40pm. Done with chores. Damn this really took a while. Let me finally start.

Spiking NNs, spiking NNs, spiking NNs...

Spiking NNs are the future I need to aim for.

1:45pm. With every day that passes the future that I seek is coming closer to reality.

I do not know how things will go with spiking NNs...there will be multiple competing companies. For myself, once I get v0.2 to a marketable state, I will peddle it around in return for a remote job. This will kill a few birds in one swoop: my money issues, allow me continue to improving Spiral, boost the whole field by relieving them of the burdens of doing everything on C under the hood, and lastly it will allow me to just sit there and continue fishing. Eventually the principles of learning will be understood and that is when I will be able to do my real work.

I really was a huge idiot back in 2018. Really, the only thing of value I got in that period was experience in programming Spiral. The rest is either basic knowledge or negatives. I have a broad, but useless base.

1:55pm. If somebody plopped the post-Singularity models down on my hard drive I could really learn so much from that. That is just daydreaming. It won't happen. 'We' need to figure it out ourselves and prove that we deserve that power.

Until I get a lead on what should be done, I might as well master the fundamentals of programming.

Every time I think I got it, that turns out to be just my ego talking.

2pm. The combination of good languages and new hardware appropriate for the domain will make things just so much easier.

One thing I really like about SNNs are that they are more than just backprop. Meaning they cover a much bigger domain that DL libraries like PyTorch can deal with.

I really have opportunity to contribute here.

2:05pm. Even if I wanted to do research on my own, CPUs just won't cut it. And GPUs have proven their crapiness quite a few times already.

I absolutely need those chips if I am going to make any progress on ML myself.

And before that, I need the completed Spiral.

Just how long am I going to work on this shit? Let me compete the damn partial evaluator already.

And then I will do typechecking.

Time consuming or not, typechecking will a major research subject to me. When I master unification I will have uncovered an entirely new domain that I haven't had before.

So I should look forward to it at least a little.

Rather than being a tedious chore, maybe it will turn out to be an incredible breath of fresh air in my programming.

Let me finally start this thing. Finish the peval today should not be hard. Let me do it in the next few hours.

Then comes the codegen. That should be easy as well."