"9:15am. I am up, and I've realized that I've messed up the renaming everywhere. The hash consing that I used to do was right, what I have now is just spew. Damn.
I need to go over it once more.

9:20am. For some reason the Internet is down.

```
let data_to_ty x =
    let m = Dictionary(HashIdentity.Reference)
    let rec f x =
        let memoize f = memoize m f x
        let memoize_rec e ret f = memoize_rec m e ret f x
        match x with
        | TyPair(a,b) -> memoize (fun _ -> PairT(f a, f b))
        | TyKeyword(a,b) -> memoize (fun _ -> KeywordT(a, Array.map f b))
        | TyFunction(a,b,c,d,e,z) -> memoize_rec e (fun e' -> FunctionT(a,b,c,d,e',z)) f
        | TyRecord l -> memoize (fun _ -> RecordT(Map.map (fun _ -> f) l))
        | TyV(T(_,ty) as t) -> ty
        | TyLit x -> value_to_ty x
        | TyB -> BT
        | TyRV _ | TyRR _ -> failwith "Compiler error: TyRV and TyRR should not appear here."
    f x
```

Uf, I did not do this right.

```
| TyFunction(a,b,c,d,e,z) -> memoize_rec e (fun e' -> FunctionT(a,b,c,d,e',z)) f
```

Yeah, sure this will do the job, but it will also leave the equality everywhere else to diverge.

9:30am. Ah, for fuck's sake...

10am. Ok, I am almost done chilling. I have no choice at this point, I am going to have to redo the renaming functions.

10:10am. I need to think a bit how to do it though.

I do not think I should do global hash consing for this. Instead it should be local.

But that will just make things even more complicated.

10:30am. This is really beyond annoying.

Forget my plans, I need to rethink this from the ground up.

It just goes to show how difficult this really is.

Now that I have recursive functions, I am making mistakes all over the place.

11:10am. I am still thinking about it full tilt.

11:35am. Ok, I sort of got it.

The conversion of data_to_rdata will have two parts...first I will convert all the recursive functions to TyRRs. I am going to have to use the nested dictionary trick I use for cse whenever I do a recursive function.

The the second part is straightforward hash consing.

11:45am. There is just one thing that bothers me here.

I've just realized that the most effective way to do renaming is in the global order of appearance. But the way prepass orders the variables in the environment is by their appearance in the function body.

This is really a problem for join point specialization.

11:55am. Still what about the order in layout types?

12:05pm. Well, nevermind those. I do not require the same degree of precision in layout types as I do in regular stuff.

12:30pm. Programming can be quite stressful. I am still thinking about it.

```
| TyFunction(a,b,c,d,e,z) -> memoize_rec e (fun e' -> FunctionT(a,b,c,d,e',z)) f
```

Actually this is fine as `FunctionT` cannot appear in a variable.

Later I might have to redo it if I introduce recursive type functions, but I can leave that for much later.

12:35pm. I am going to leave optimizing the prepass for later. Right now, I am just going to do hash consing of data properly."