"1:20pm. It is just me vs the recursive cycles here, and the recursive cycles are killing me.
1:25pm. I am confused. At this point I've gone so far into the rabbit hole that I've forgotten what the hell was I doing in the first place.

I wanted to figure out a way to hash cons the individual pieces, right? In order to make the join point comparisons faster?

I woke up all of a sudden realizing that cycles in `FunctionT` are a problem, except they aren't.

Ah, yes, ah yes...

I realized something horrible.

```
inl rec f a b c = ...
inl q = f a
inl w = f a
join q b c, w b c
```

What I realized in the morning is that my assumption that functions cannot be duplicates is broken.

I realized this about pairs, keywords and records earlier. But now it turns out that the case is that the same specialization trick will fail for functions as well.

So renaming is broken.

I need to do hash consing.

But the cycles, really, really get in the way here.

3:20pm. I am really being screwed here.

I can't find a way to code myself out of this mess.

I really might have to go to the v0.09 of doing recursion without necessarily having cycles directly in the environment.

3:25pm. The disadvantage of this is at runtime though.

I've sort of made it a principle not to sacrifice runtime in favor of the compile time, but these cycles are a big deal.

3:45pm. I am still thinking about it.

I've sort of decided - I am going to avoid having cycles in the environment.

They just have too many disadvantages during join point specialization that I have to do this.

Even if means having to do a naive scheme that has disadvantages at runtime, I have to do this.

3:50pm. Previously, I've sort of taken it for granted that it is not possible to do the environments indirectly.

But is that really true?

4:05pm. This thing has pretty much ruined my entire day. I should be starting work on the codegen, instead I am bashing my head against the wall here against this issue that should have been solved long ago.

Let me step away from the screen here. I am going to take a nap instead of just spinning my wheels here."