"10:50am. Uf, I got up late. I might as well have breakfast now then.
11:50am. Done with that. I do not feel like chores straight away. I am still groggy, so let me slack just a bit.

11:55am. Just one ep of Railgun and then I will do the chores. After that comes the codegen.

12:25pm. Chores.

1pm. Now finally it is time to do some programming. I have 5 good hours ahead of me. Let me put them to good use.

Focus me, focus.

Just think...once I finish the codegen, and then get the language running, I will essentially have what I had back in 2018 except without any flaws. No more compilation performance issues.

That is something.

At that point should I want it, I could retrace my old steps and resume where I left off in 2018.

Except I want to go a lot further this time and I will.

It is time for mainstreaming of Spiral's features. I will give it a honest shot at showing the world its full potential. I won't stop at having it be a power tool for experts.

1:05pm. Everything. This time there will be no more compromises in the ability to reason. This time I will have everything.

The perfect language. The proper substrate. The needed algorithms.

The 20s are going to have my everything.

1:15pm. Yeah, my energy is low as I am dealing with the redesign bullshit instead of the kind of programming that I want. When I start working on typechecking, and later do the plugin, it will go up.

I was wrong in 2018. It is not my job to do ML at this stage. Rather I need to do basic research. That means finding the use for spiking nets and such. I won't get anywhere until I figure out the improved data structures that could be used in place of matrices of floats.

I will make it my principle to look for that. I won't move to doing real world ML until this has been done.

That will be one way to atone for my past misdeeds.

1:20pm. If I could just get a little bit more understanding, I could finish the deal myself without relying on the rest of the field so much.

The hints are already all there, just the program specifics are the ones which are being elusive.

1:25pm. I need to bear the burden of not knowing until it happens.

First of all, let me get the things out of the way. I do not want to spend any more time thinking about the partial evaluator or the codegen or whatever.

So let me get that finally out of the way so that I can start doing more important work.

1:30pm.

```
exception CodegenError of string
let raise_codegen_error x = raise (CodegenError x)
```

I do not require codegen errors anymore as I no longer have the danger of type spilling on the value level.

But I will leave it in just in case.

No need to get rid of this.

Now...

Yesterday I removed the printing of the join point cases. To start things off today I will put `RJPToStack` and `RJPToHeap` back in.

1:55pm.

```
        | TyLet(a,trace,x) ->
            try
                let vars = data_free_vars a |> tytags_comma d
                let print_let b = d.Statement(sprintf "let %s = %s" vars b)
                match x with
                | TyOp(RJPToStack,[|b|]) ->
                    match data_free_vars b with
                    | [||] -> "() // unit stack layout type"
                    | free_vars ->
                        let tag = d.types.Tag (data_to_ty a)
                        let b = tytags_comma' d free_vars
                        sprintf "SpiralType%i %s" tag b
                    |> print_let
                | TyOp(RJPToHeap,[|b|]) ->
                    match data_free_vars b with
                    | [||] -> "() // unit heap layout type"
                    | free_vars ->
                        free_vars
                        |> Array.mapi (fun i x -> sprintf "subvar_%i = %s" i (tytag' d x))
                        |> String.concat "; "
                        |> sprintf "{%s}"
                    |> print_let
                | TyOp(RJPToNone,[|b|]) ->
                    match b with
                    | TyV(T(tag,RJPT(_,b))) ->
                        rdata_free_vars b
                        |> Array.map (fun (tag',_) -> sprintf "var_%i.subvar_%i" tag tag')
                        |> String.concat ", "
                        |> print_let
                    | _ -> failwith "impossible"
                | TyJoinPoint _ | TyOp _ -> print_let (op d x)
                | _ -> d.Statement(sprintf "let %s =" vars); op d.Indent x |> ignore
            with :? CodegenError as x -> raise (CodegenErrorWithPos(trace,x.Data0))
        | TyLocalReturnOp(trace,x) ->
            try match op d x with null -> () | x -> d.Statement(x)
            with :? CodegenError as x -> raise (CodegenErrorWithPos(trace,x.Data0))
        | TyLocalReturnData(x,trace) ->
            try d.Statement(typed_data d x)
            with :? CodegenError as x -> raise (CodegenErrorWithPos(trace,x.Data0))
```

I am starting to get the feel back. I've realized that the stack does in fact need the return type in the option after all.

Because otherwise if it is not a part of the let statement it will fail. Whops.

I definitely botched that one.

...Ok, let me go back to the partial evaluator...

2:05pm.

```
let rjp_to_some layout dict (d: LangEnv) = function
    | TyV(T(_,RJPT(t,l))) as x when t = layout -> x
    | x ->
        let x = rjp_to_none d x
        let op = match layout with RJPStack -> RJPToStack | RJPHeap -> RJPToHeap
        let key = op,[|x|]
        match cse_tryfind d key with
        | Some x -> x
        | None ->
            let ret = ty_to_data d.i (RJPT(layout,data_to_rdata dict x))
            d.seq.Add(TyLet(ret,d.trace,TyOp(op,[|x;ret|])))
            cse_add d key ret
            ret
```

Let me do this shit.

Now...

```
        | RJPToStack, [|b;a|] ->
            match data_free_vars b with
            | [||] -> "() // unit stack layout type"
            | free_vars ->
                let tag = d.types.Tag (data_to_ty a)
                let b = tytags_comma' d free_vars
                sprintf "SpiralType%i %s" tag b
        | RJPToHeap,[|b;a|] ->
            match data_free_vars b with
            | [||] -> "() // unit heap layout type"
            | free_vars ->
                free_vars
                |> Array.mapi (fun i x -> sprintf "subvar_%i = %s" i (tytag' d x))
                |> String.concat "; "
                |> sprintf "{%s}"
        | RJPToNone,[|b|] ->
            match b with
            | TyV(T(tag,RJPT(_,b))) ->
                rdata_free_vars b
                |> Array.map (fun (tag',_) -> sprintf "var_%i.subvar_%i" tag tag')
                |> String.concat ", "
            | _ -> failwith "impossible"
```

Plugged it in. Great.

2:10pm. Since I haven't done it yet, let me commit here.

What comes next is the last bit of the codegen.

But besides that, what is really troubling me is that in the previous version of Spiral I had that thing where I convert unit types to `TyT`s.

Now I do not have that anymore, so things are more complicated. I need to that bit of reasoning from the ground up.

But first let me get rid of the red.

2:20pm. Focus me, focus."