"4:25pm. Done with the break. Let me finally resume. I need to start thinking how to compile. It is time to get this party started.
There are 160 lines in `SpiralCompile` and I need to start going through them.

Like with the codegen, let me first break the ice and then I will do it properly.

4:35pm. Ok...

First of all, let me put all the types into a single file. What I have now is unmanageable.

I read that article on Haskell urging devs not to centralize the types, but this is not necessarily good advice for a compiler.

Let me fix it.

4:40pm. I am doing it now and I am not getting the feeling that I am making the right choice here.

4:55pm. I've decided no to do it after all. The article was right. I'll just open everything only for the types.

F# isn't autistic about this like Agda or Haskell.

5pm. I've thought about it for a while.

Basically, everything in the old `SpiralCompile` is obsolete. I have nothing to work with here. I need to redo it completely.

It does things like raise an error on unbound variables and deals with module bindings and opens which are now the provenance of the prepass.

5:05pm. Ah, I sure do not feel like it.

But I must.

Let me make the `compile_module` function for starters.

...Actually, I need to take a break here."