"9:40am. Ok, enough chilling. I've been doing that for nearly an hour. It is time to get something done.
9:45am. Let me open the IDE...

```
let test41: SpiralModule =
    {
    name="test40"
    prerequisites=[]
    description="Does the let statement work?"
    code=
    """
inl main _ =
    let f (x : i64) = x
    f 3
    """
    }
```

This just gets me `Box not implemented`.

```
| Op(op,_) -> raise_type_error d <| sprintf "Compiler error: %A not implemented" op
```

This is strange. At least when annotation is given I'd expect it to go through a different channel.

```
    | AnnotTest(do_boxing,ty,expr,on_succ,on_fail) ->
        let a = v expr
        let b = tev d ty
        if do_boxing then ev (push_value_var (box' d (a,b)) d) on_succ
        else if data_to_ty a = b then ev d on_succ else ev d on_fail
```

I mean this thing here.

Actually, do I even need the regular box?

Well...

Actually, I should make `ty` optional here. That way the typechecker will know where it has to put things.

10:10am. I am thinking here.

All this is so complicated.

Just how many different ways of boxing can there be?

If I am going to box, the most sensible thing would be to put the annotation directly in the function. That way, the typechecker will later know that it has to propagate this information and put it in the right place.

10:45am. https://thegradient.pub/independently-reproducible-machine-learning/

Let me read this.

I am still thinking about the whole issue of annotations vs boxing.

11am. Still thinking about it...

11:35am. In the bathroom...on this chair...I am still thinking about it.

11:55am. I can't believe I am completely stuck on this. It is not even closure conversion that is the problem, but how to deal with annotations instead.

I put in let statements haphazardly, saying the typechecker will take care of it, but now it is the time to really think about it.

What I am thinking about now is skipping the closure conversion thing and moving straight onto typechecking. That will clear things up for me.

12pm. Let me stop here. I did not do any programming to speak of despite having a head start.

Damn it, I really did want to make progress on closure conversion this morning too.

Nevermind. Forget that.

That and macros will be done in due time.

What I am going to do now is start the second part of the language. The parts up to now were only there to finally deal with the compilation issues. It is the top down typing that will really be Spiral's wings. It is this what will allow me to fight through the cruel world of computer science.

12:10pm. Ah damn it, now I am reading shitty ML sub threads. Let me have breakfast here.

I hadn't planned on starting typechecking right away so I've been caught flatfooted.

I am going to think about this for a bit and then just get on with it.

I'll do closure conversion while working on part two.

It is important that I get this done. Typechecking via constraint solving is something new even to me, and it will be the first time I will be working on this. And while I am doing this, I will also be ironing out the syntax so that the two pieces integrate nicely.

It is only after I've done this that I can say that the new language is complete.

This will be the real deal. The one true test of skill in PL development.

After I do this, I really will have done it all.

After it is done, all that will remain is for me to fish for the right ML algorithms."