"5:10pm. Ok, I've somewhat fortified my mind. I've been in bed all this time thinking. I'd be lying if I said I was thinking about programming only, but going through blocks like these takes time.
Ok, first thing's first.

The part one of the language is nearly done. The things that are missing: union types, closures, macros are minor challenges that if wasn't intending to move to typechecking I would finish swiftly in a day or two. So that is what I mean by nearly done.

5:20pm. Right now I've been approaching this problem from the perspective of extending the work that I've done.

5:25pm. But that is not the right way to go about this.

Instead I need to fully commit to the top-down aspect. I won't be able to do the integration until I master it.

For this kind of planning to be done, it is not enough to have a rough conceptual understanding. My mastery of the subject has to be complete.

5:30pm. So here is what I will do. I will literally abandon the v0.2 stuff that I've done up to now. I can lie as it is for a few months.

Instead what I will do is start a new project, that is focused exclusively on the typechecking aspects of Spiral.

My goal will just be to do the typechecking, and alongside that I will begin work on the VSCode plugin.

When that is done, I will bring in the constraints and the partial evaluator.

5:35pm. It will be much easier to develop the new half in isolation and connect it later rather than do it all simultaneously.

I want to fully dedicate myself mentally to this next part without necessarily burdening myself with thoughts of what the prepass or the partial evaluator should be doing.

5:40pm. It is much earlier than I thought it would be, but the gong of destiny has been struck. It is time to begin the second fateful part of Spiral.

The big lesson from all of this is that even though partial evaluation can be used to do anything, it is not necessarily the best idea to actually do so.

The parallels with deep learning aren't escaping me.

Backprop is a key thing that can potentially do anything, but cannot scale all the way. Naive ways of doing regularization: L2 norm, weight norm, layer norm...aren't enough to deal with its issues. Neither can architectural tricks.

There is definitely something missing before the full picture can be completed.

In case of Spiral, I can see the issues and the problem clearly, but in the case of deep learning only the issues are clear. The neural equivalent of unification is shrouded in absolute darkness.

When the veil gets lifted, I will be able to move on to the next step and fulfill the old promises of the last decade.

5:50pm. Well, I'll take it easy from here.

Typechecking and the plugin sound like a big deal, but compared to the things I've done so far, they should not be too much.

Contextually, it seems partial evaluation is a more advanced subject than typechecking so I should be safe as long as I take it one at a time. Rather than let myself be crushed by the immensity of work I've yet to do, I need to look at the lighter aspects of the work that is to come.

For the first time in my life, I will be writing top-down Spiral code. If that is not something to look forward to, I do not know what is."