"4:10pm. Focus me, focus. Read the documentation, not Twitter posts.
4:25pm. Focus, focus...

There is a bunch in the documentation on language plugins. That is what i need to study.

4:50pm. https://code.visualstudio.com/api/language-extensions/programmatic-language-features
"The process seems more complicated, but it provides two major benefits:

The Language Server can be written in any language
The Language Server can be reused to provide smart editing features for multiple editors"

So LSP is what I need.

4:55pm. I've come to a design decision.

I think that for the real segment, I will just disallow let statements due to their awkwardness there.

There is just no need to bother with them there.

5:05pm. https://code.visualstudio.com/api/language-extensions/language-server-extension-guide
"However, while implementing support for language features in VS Code, we found three common problems:

First, Language Servers are usually implemented in their native programming languages, and that presents a challenge in integrating them with VS Code, which has a Node.js runtime."

Yeah, this is the thing I've been wondering all this time. So far all the examples I've seen have been in Typescript. I certanly cannot reimplement all the necessary stuff in TS.

5:30pm. Done with lunch. Let me resume this.

6:05pm. Usually at this point I would stop for the day, but since I slacked so much earlier in the day, I will just continue for a while longer.

https://martinbjorkstrom.com/posts/2018-11-29-creating-a-language-server

This is what I am reading now.

Creating a LSP will really be a pain in the ass it seems, at least in the beginning.

http://www.omnisharp.net/
"OmniSharp is a set of tooling, editor integrations and libraries that together create an ecosystem that allows you to have a great programming experience no matter what your editor and operating system of choice may be."

I guess I will try installing this.

https://www.nuget.org/packages/OmniSharp.Extensions.LanguageServer/
"You can use this package to create a language server, that will be consumed by many clients"

Definitely sounds like what I need.

6:40pm. https://github.com/fsprojects/fsharp-language-server/blob/master/src/FSharpLanguageServer/Program.fs

I am currently looking at the stuff here. I hadn't realized LSP communicates using JSON.

6:50pm. https://github.com/fsprojects/fsharp-language-server

Ok, this is interesting. I might want to clone this and use it as a template.

I'll leave studying it as plan B.

Plan A will be to go with...

https://martinbjorkstrom.com/posts/2018-11-29-creating-a-language-server

I will study this in the depth and try to get something basic running tomorrow. Something that highlights some keywords in red, or basic autocomplte a -> abc. Something dumb like that to serve as hello world icebreaker.

6:55pm. I need to do this. Eventually I am going to figure out how to deal with all the type system features that Spiral requires, so it makes sense to start from the other end. Typechecking is not a big challenge by itself.

Unless I constantly want to jump back and forth between scripts like I did in v0.09, I need to write my own tooling. I can't leave this for later or to some adoring fan of Spiral. I can't leave the tough and dirty stuff to others. I need to do it myself.

7pm. The language that I am creating here is my own power that will sustain me.

It is a much bigger and better fishing net than v0.09.

7:05pm. All in all, I am glad I decided to go down this route. The question of designing let statements gave me a lot of consternation, but it seems leaving it out is the right answer.

The vast majority of Spiral code is going to be in the inferred segment, so I should kill myself trying to put special syntax into the thing.

I already mastered partial evaluation aspects with 0.09. Once I do the tooling and the typechecking, I will have mastered the low style of programming. The low style is the best style.

The higher (dependently typed) style is more expressive, but dependent types are a productivity wreck. They also necessitate dynamic runtimes so that makes the kind of use cases I am trying to hit with Spiral out of reach.

...Ah, damn it. I should studied physics in 2019 instead of type theory.

...I do not really mean it. I really did need to go through this once to form a proper view on math itself.

Well, things will go quietly.

It is really a pity RL is broken, but that will give me time to give Spiral another chance. This time I will do things right.

7:20pm. I won't abandon it again. This time for sure, I will turn it into power.

Without a doubt, better, newer hardware will uncover new vistas of research just like GPUs have.

I was just a bystander in the 10s. I was there when Hinton announced his breakthrough results in the Coursera course back in 2012. If possible, the next time that happens I want to be in the thick of things having anticipated what is to come.

The story of AI won't end here. It won't petter out here only to resurface 50-100 years from now.

The story of the coming Singularity will play out right here and in my lifetime. The true players are those actively trying to grasp the future.

They are never those who are making noise and just talking about it."