"9am. https://www.reddit.com/r/MachineLearning/comments/f1tuv0/r_turingnlg_a_17billionparameter_language_model/ https://www.reddit.com/r/MachineLearning/comments/f1tzqe/n_microsoft_researchers_and_engineers_release/ https://www.reddit.com/r/MachineLearning/comments/f1vqvh/compressive_transformers_for_longrange_sequence/ https://www.reddit.com/r/MachineLearning/comments/f1tutb/research_intuitive_gaussian_mixture_estimation/ An Intuitive Introduction to the EM Algorithm for Gaussian Mixtures
9:05am. There is a lot of interesting stuff on the ML sub today.

https://news.ycombinator.com/item?id=22288735
JuliaLang: The Ingredients for a Composable Programming Language

https://news.ycombinator.com/item?id=22281561
Nontrivially fillable gaps in published proofs of major theorems

These two threads on Julia and math respectively were of interest to me yesterday.

"What does this have to do with garbage collection?
If your concern is about latency in real time systems, Garbage collection isn't the problem, heap allocating memory is. Julia makes it easy to never allocate anything on the heap (unlike most garbage collected langauges).

Check out this talk this discusses this in detail in the context of robotics: https://www.youtube.com/watch?v=dmWQtI3DFFo"

In particular this post caught my attention and I decided yesterday that I am going to watch this talk today.

9:35am. Ah, this will take a while. I have barely even started this.

10am. https://arxiv.org/pdf/1911.05507.pdf
Compressive transformers for long-range sequence modelling

10:10am. 6/18. The Enwiki benchmark really has been crushed by these transformers hasn't it?

It is a pity I am out of it right now, but it does seem these architectures will be here to stay.

I really do not understand them just yet. I'll leave this task for later.

"We break perplexity down by word frequency in Table 7 and see the Compressive Transformer makes only a small modelling improvement for frequent words (2.6% over the TransformerXL baseline) but obtains a much larger improvement of ≈ 20% for infrequent words. Furthermore, we see 10X improvement in modelling rare words over the prior state-of-the-art LSTM language model published in 2018 — which demonstrates the rate of progress in this area"

10:40am. Finally got through all the ML sub links. Let me go through the Julia talk.

11:15am. https://youtu.be/dmWQtI3DFFo?t=1624

Finally, he gets to the point.

11:30am. https://youtu.be/dmWQtI3DFFo?t=2271
"A macro that turns allocation into error would make debugging much easier."

He talks a lot about how great static arrays and stack allocated structs are for performance.

11:40am. Ok, if it is just this, I can beat Julia. Spiral definitely does make it easy to know heap from stack allocations.

I'll take a break here. The course will have to be done in the rest of the day.

It is time to rest for a bit here.

I do not think it will take me that long go through the course as it is mostly text. 2-3 hours will do and after that...well, I think I'll have to go through the TS documentation itself.

One thing I am still missing is that I haven't been able to see the complete configuration, so I still do not know how to compile apart from the command line. I really do not want to spend money on that course, mostly because I do not have it.

Well, if I look around, I am sure I will find the solution.

11:45am. Taking it easy for a bit, while I get my TS knowledge up and running should be fine."