"1:10pm. Wow, I am wasting time. Well, let me do it for a bit longer, do the chores and then I will start this thing.
2pm. Breakfast done, chores done. It is time to start studying Typescript. Let me go through that course. Then I will stop wasting time and try that demo plugin in Typescript itself.

I want to get this done rather than getting buried in APIs.

I am a programmer, so what I should be doing is programming. Studies are for the students.

Now where was I?

2:15pm. https://courses.edx.org/courses/course-v1:Microsoft+DEV273x+1T2019a/courseware/dbe869f1929d4e46b3b03e60361b9a24/7d6a7660b2544bfabc5d1bd7987aec44/?child=first

This is really boring. I should just skim this and learn as I go along.

2:35pm. https://courses.edx.org/courses/course-v1:Microsoft+DEV273x+1T2019a/courseware/dbe869f1929d4e46b3b03e60361b9a24/7d6a7660b2544bfabc5d1bd7987aec44/?child=first

This is super boring. Let me watch the video.

3:15pm. This is ultra super boring!!!

Let me finish module 2 and I will reconsider my path.

3:30pm. Let me go through section 3.

This is such a waste of time, but on the other hand, it is super basic so I have to do it to get a passing familiarity with Typescript.

3:55pm. I am taking a little break right now. The tedium of this is really doing me in.

4:30pm. Oh, TS compiler will automatically do the import statements as well. That is nice.

4:55pm. Finally done with section 3. Yeah, the tedium of going through this shit is definitely high. It seems easy tasks have a way of making up for it by being boring.

5pm. Ok, ok...

Let me just go through this last thing and then I'll quickly peek at the TS documentation.

5:10pm. Uf, let me take a break here. I need a break.

6:05pm. Done with lunch. Let me finish up the section 4 videos and then I am done for the day.

6:30pm. https://www.typescriptlang.org/docs/handbook/decorators.html

This OO lingo for what should be simple function composition is getting to me.

https://www.typescriptlang.org/docs/home.html

Instead of trying to read through that course, I should have gone for the documentation instead.

6:35pm. But honestly there is no way I can do that right now. I am dead here.

This was so tedious, I was barely paying attention. Yet, I did my best. Tomorrow I will give the documentation a brush and after that I will jump straight into programming.

6:45pm. I need to start playing around with some basic plugins. I need to start establishing control over the editor. Once I do that, everything will be much easier.

7:05pm. When things get hard, it is easy to forget what your goals are. Constant remainders are needed to remain on course."